+++ pac4j-saml/src/main/java/org/pac4j/saml/config/SAML2Configuration.java
@@ -1,15 +1,9 @@
 package org.pac4j.saml.config;
 
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-import lombok.With;
+import lombok.*;
 import lombok.experimental.Accessors;
 import lombok.extern.slf4j.Slf4j;
-import lombok.val;
 import net.shibboleth.shared.net.URIComparator;
 import net.shibboleth.shared.net.impl.BasicURLComparator;
 import org.apache.commons.lang3.StringUtils;
@@ -21,8 +15,6 @@ import org.opensaml.xmlsec.config.impl.DefaultSecurityConfigurationBootstrap;
 import org.opensaml.xmlsec.signature.support.SignatureConstants;
 import org.pac4j.core.client.config.BaseClientConfiguration;
 import org.pac4j.core.exception.TechnicalException;
-import org.pac4j.core.logout.handler.DefaultSessionLogoutHandler;
-import org.pac4j.core.logout.handler.SessionLogoutHandler;
 import org.pac4j.core.profile.converter.AttributeConverter;
 import org.pac4j.core.resource.SpringResourceHelper;
 import org.pac4j.core.util.CommonHelper;
@@ -30,16 +22,7 @@ import org.pac4j.core.util.Pac4jConstants;
 import org.pac4j.saml.crypto.CredentialProvider;
 import org.pac4j.saml.crypto.KeyStoreCredentialProvider;
 import org.pac4j.saml.exceptions.SAMLException;
-import org.pac4j.saml.metadata.BaseSAML2MetadataGenerator;
-import org.pac4j.saml.metadata.SAML2FileSystemMetadataGenerator;
-import org.pac4j.saml.metadata.SAML2HttpUrlMetadataGenerator;
-import org.pac4j.saml.metadata.SAML2IdentityProviderMetadataResolver;
-import org.pac4j.saml.metadata.SAML2MetadataContactPerson;
-import org.pac4j.saml.metadata.SAML2MetadataGenerator;
-import org.pac4j.saml.metadata.SAML2MetadataResolver;
-import org.pac4j.saml.metadata.SAML2MetadataSigner;
-import org.pac4j.saml.metadata.SAML2MetadataUIInfo;
-import org.pac4j.saml.metadata.SAML2ServiceProviderRequestedAttribute;
+import org.pac4j.saml.metadata.*;
 import org.pac4j.saml.metadata.keystore.SAML2FileSystemKeystoreGenerator;
 import org.pac4j.saml.metadata.keystore.SAML2HttpUrlKeystoreGenerator;
 import org.pac4j.saml.metadata.keystore.SAML2KeystoreGenerator;
@@ -57,14 +40,7 @@ import javax.net.ssl.HostnameVerifier;
 import javax.net.ssl.SSLSocketFactory;
 import java.net.URL;
 import java.time.Period;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.LinkedHashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.ServiceLoader;
+import java.util.*;
 import java.util.function.Supplier;
 
 /**
@@ -203,8 +179,6 @@ public class SAML2Configuration extends BaseClientConfiguration {
     private Map<String, String> mappedAttributes = new LinkedHashMap<>();
 
     private URIComparator uriComparator = new BasicURLComparator();
-
-    private SessionLogoutHandler sessionLogoutHandler;
 
     private String postLogoutURL;
 
@@ -355,10 +329,6 @@ public class SAML2Configuration extends BaseClientConfiguration {
             keystoreGenerator.generate();
         }
 
-        if (sessionLogoutHandler == null) {
-            sessionLogoutHandler = new DefaultSessionLogoutHandler();
-        }
-
         initSignatureSigningConfiguration();
     }
 
@@ -466,17 +436,6 @@ public class SAML2Configuration extends BaseClientConfiguration {
      */
     public void setServiceProviderMetadataPath(final String path) {
         this.serviceProviderMetadataResource = SpringResourceHelper.buildResourceFromPath(path);
-    }
-
-    /**
-     * <p>findSessionLogoutHandler.</p>
-     *
-     * @return a {@link SessionLogoutHandler} object
-     */
-    public SessionLogoutHandler findSessionLogoutHandler() {
-        init();
-
-        return sessionLogoutHandler;
     }
 
     private void initSignatureSigningConfiguration() {